+++ src/bump_string.rs
@@ -7,8 +7,7 @@ use core::{
 };
 
 use crate::{
-    error_behavior_generic_methods_allocation_failure, polyfill, BaseAllocator, BumpBox, BumpScope, BumpVec, ErrorBehavior,
-    FromUtf8Error, GuaranteedAllocatedStats, MinimumAlignment, Stats, SupportedMinimumAlignment,
+    error_behavior_generic_methods_allocation_failure, polyfill, BaseAllocator, BumpBox, BumpScope, BumpVec, ErrorBehavior, FixedBumpString, FromUtf8Error, GuaranteedAllocatedStats, MinimumAlignment, Stats, SupportedMinimumAlignment
 };
 
 /// This is like [`format!`] but allocates inside a `Bump` or `BumpScope`, returning a [`BumpString`].
@@ -550,6 +549,16 @@ where
     /// ```
     pub fn shrink_to_fit(&mut self) {
         self.vec.shrink_to_fit();
+    }
+
+    /// Turns this `BumpString` into a `FixedBumpString`.
+    ///
+    /// You may want to call [`shrink_to_fit`](Self::shrink_to_fit) before this, so the unused capacity does not take up space.
+    #[must_use]
+    #[inline(always)]
+    pub fn into_fixed_string(self) -> FixedBumpString<'a> {
+        let vec = self.vec.into_fixed_vec();
+        unsafe { FixedBumpString::from_utf8_unchecked(vec) }
     }
 
     /// Converts a `BumpString` into a `BumpBox<str>`.